[ci] fix failing workflows: SWIG download hardening, lint, typos, cpplint
- SWIG (windows): pin swigwin 4.2.1 (sourceforge 'latest/download' serves an
  HTML rate-limit page intermittently), retry up to 5x and validate the zip
  before extracting.
- ruff: per-file-ignores for benchmarks/* CLI scripts (D103/T201/PLC0415),
  fix I001/B905/PLW1510/D401 properly; ruff-format 4 benchmark files.
- typos: fmt(nd=) -> digits, reword 'ANDed' comment, ignore hex commit hashes
  in .typos.toml (ROADMAP.md references them).
- cpplint: include order + missing includes in 5 CUDA files; %lld casts ->
  PRId64 in the selective-growth stats fprintf.

pre-commit run --all-files is green locally at ruff 0.14.10 / typos 1.40.0.

+++ src/treelearner/cuda/cuda_single_gpu_tree_learner.cpp
@@ -18,6 +18,8 @@
 #include <algorithm>
 #include <array>
 #include <chrono>
+#include <cinttypes>
+#include <string>
 #include <cstdio>
 #include <cstdlib>
 #include <memory>
@@ -1626,11 +1628,11 @@ void CUDASingleGPUTreeLearner::SelectiveFinalize(CUDATree* tree) {
   static const bool hybrid_debug = std::getenv("EXABOOST_HYBRID_DEBUG") != nullptr;
   if (hybrid_debug) {
     // stderr (not the logger): churn statistics must be visible under verbose=-1
-    fprintf(stderr, "[selective] trees=%lld leaves=%d cumulative: applied=%lld displaced=%lld levels=%lld\n",
-            static_cast<long long>(sel_stat_trees_), final_num_leaves,
-            static_cast<long long>(sel_stat_applied_),
-            static_cast<long long>(sel_stat_displaced_),
-            static_cast<long long>(sel_stat_levels_));
+    fprintf(stderr, "[selective] trees=%" PRId64 " leaves=%d cumulative: applied=%" PRId64 " displaced=%" PRId64 " levels=%" PRId64 "\n",
+            static_cast<int64_t>(sel_stat_trees_), final_num_leaves,
+            static_cast<int64_t>(sel_stat_applied_),
+            static_cast<int64_t>(sel_stat_displaced_),
+            static_cast<int64_t>(sel_stat_levels_));
   }
 }
